services/verification/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sentence-transformers==3.3.1
numpy==2.2.1
redis==5.2.0
python-dotenv==1.0.1
httpx==0.28.1
langdetect==1.0.9
